refactor commandcontroller -> command 용도 분리
