+++ src/main/java/org/prgrms/application/Command.java
@@ -1,5 +1,6 @@
-package org.prgrms.application.controller;
+package org.prgrms.application;
 
+import org.prgrms.application.controller.VoucherController;
 import org.prgrms.application.view.InputView;
 import org.prgrms.application.view.OutputView;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,16 +8,18 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
 @Component
-public class CommandController implements CommandLineRunner{
+public class Command implements CommandLineRunner{
 
     private final InputView inputView;
     private final OutputView outputView;
+    private final VoucherController voucherController;
     private boolean isRunning = true;
 
     @Autowired
-    public CommandController(InputView inputView, OutputView outputView) {
+    public Command(InputView inputView, OutputView outputView, VoucherController voucherController) {
         this.inputView = inputView;
         this.outputView = outputView;
+        this.voucherController = voucherController;
     }
 
     @Override
@@ -33,7 +36,7 @@ public class CommandController implements CommandLineRunner{
         switch (commandType){
             case CREATE -> {
                 String selection = inputView.selectVoucherType();
-                V
+
             }
 
 //            case LIST -> {
