+++ src/main/java/org/prgrms/application/controller/VoucherController.java
@@ -1,0 +1,14 @@
+package org.prgrms.application.controller;
+
+import org.prgrms.application.service.VoucherService;
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class VoucherController {
+
+    private VoucherService voucherService;
+
+//    public createVoucher(VoucherService voucherService){
+//        voucherService.createVoucher();
+//    }
+}
